+++ velesresearch/wrappers.py
@@ -44,6 +44,7 @@ def survey(
     mode: str = "edit",
     navigateToUrl: str | None = None,
     navigateToUrlOnCondition: list[dict] | None = None,
+    navigationButtonsLocation: str = "bottom",
     numberOfGroups: int = 1,
     pageNextText: str | None = None,
     pagePrevText: str | None = None,
@@ -63,7 +64,7 @@ def survey(
     requiredMark: str = "*",
     scoresSuffix: str = "_score",
     showCompletePage: bool = True,
-    showNavigationButtons: str = "bottom",
+    showNavigationButtons: bool = True,
     showPageNumbers: bool | None = None,
     showPageTitles: bool = True,
     showPrevButton: bool = True,
@@ -132,6 +133,7 @@ def survey(
         mode (str): The mode of the survey. Can be 'edit' (can be filled), 'display' (read-only).
         navigateToUrl (str | None): URL to navigate to after the survey is completed.
         navigateToUrlOnCondition (list[dict] | None): URL to navigate to after the survey is completed if the condition is met. List of dictionaries with keys `expression` and `url` keys.
+        navigationButtonsLocation (str): The location of the navigation buttons. Can be 'bottom' (default), 'top', 'topBottom'.
         pageNextText (str | None): Text for the 'Next' button.
         pagePrevText (str | None): Text for the 'Previous' button.
         previewText (str | None): Text for the 'Preview' button if `showPreviewBeforeComplete=True`.
@@ -149,7 +151,7 @@ def survey(
         requiredMark (str): The text denoting the required questions. Default is '*'.
         scoresSuffix (str): The suffix of the score column if `addScoreToResults=True`. Default is '_score'.
         showCompletePage (bool): Whether to show the completed page. Default is True.
-        showNavigationButtons (str): The location of the navigation buttons. Can be 'bottom' (default), 'top', 'both', 'none'.
+        showNavigationButtons (bool): Whether to show the navigation buttons.
         showPageNumbers (bool | None): Whether to show the page numbers in the pages' titles.
         showPageTitles (bool): Whether to show the page titles. Default is True.
         showPrevButton (bool): Whether to show the 'Previous' button. Default is True.
@@ -213,6 +215,7 @@ def survey(
         "timeLimit": timeLimit,
         "timeLimitPerPage": timeLimitPerPage,
         "mode": mode,
+        "navigationButtonsLocation": navigationButtonsLocation,
         "navigateToUrl": navigateToUrl,
         "navigateToUrlOnCondition": navigateToUrlOnCondition,
         "numberOfGroups": numberOfGroups,
@@ -277,7 +280,7 @@ def page(
     isRequired: bool = False,
     maxWidth: str = "100%",
     minWidth: str = "300px",
-    navigationButtonsVisibility: str = "inherit",
+    navigationButtonsLocation: str = "bottom",
     navigationDescription: str | None = None,
     navigationTitle: str | None = None,
     questionErrorLocation: str = "default",
@@ -286,6 +289,7 @@ def page(
     readOnly: bool = False,
     requiredErrorText: str | None = None,
     requiredIf: str | None = None,
+    showNavigationButtons: bool = True,
     state: str = "default",
     timeLimit: int | None = None,
     timeMinimum: int | None = None,
@@ -309,8 +313,7 @@ def page(
         isRequired (bool): Whether the page is required (at least one question must be answered).
         maxWidth (str): Maximum width of the page in CSS units.
         minWidth (str): Minimum width of the page in CSS units.
-        navigationButtonsVisibility (str): The visibility of the navigation buttons. Can be 'inherit', 'show', 'hide'.
-        navigationDescription (str | None): Description for the page navigation.
+        navigationButtonsLocation (str): The location of the navigation buttons. Can be 'bottom' (default), 'top', 'topBottom'.
         navigationTitle (str | None): Title for the page navigation.
         questionErrorLocation (str): The location of the error text for the questions. Can be 'default', 'top', 'bottom'.
         questionOrder (str): The order of the questions. Can be 'default', 'random'.
@@ -318,6 +321,7 @@ def page(
         readOnly (bool): Whether the page is read-only.
         requiredErrorText (str | None): Error text if the required condition is not met.
         requiredIf (str | None): Expression to make the page required (at least one question must be answered).
+        showNavigationButtons (bool): Whether to show the navigation buttons.
         state (str): If the page should be collapsed or expanded. Can be 'default', 'collapsed', 'expanded'.
         timeLimit (int | None): Maximum time in seconds to finish the page.
         timeMinimum (int | None): Minimum time in seconds the user has to spend on the page. All navigation buttons are hidden during this period.
@@ -336,7 +340,7 @@ def page(
         "timeMinimum": timeMinimum,
         "maxWidth": maxWidth,
         "minWidth": minWidth,
-        "navigationButtonsVisibility": navigationButtonsVisibility,
+        "navigationButtonsLocation": navigationButtonsLocation,
         "navigationDescription": navigationDescription,
         "navigationTitle": navigationTitle,
         "questionErrorLocation": questionErrorLocation,
@@ -345,6 +349,7 @@ def page(
         "readOnly": readOnly,
         "requiredErrorText": requiredErrorText,
         "requiredIf": requiredIf,
+        "showNavigationButtons": showNavigationButtons,
         "state": state,
         "title": title,
         "visible": visible,